concorde.cli: no subcommand help for py 3.4
Due to Issue 9351 in CPython, the subparser 'set_defaults' were not being
applied.  However the implementation relied on that to provide subparser
specific help in the case of no selected subcommands (or sub-subcommands).

This commit changes 'cli.py' to uses an alternative strategy of distinct names
for 'action' variables at each level to detect unselected subcommands.