concorde: handle 404 order objects
The assumption that an order object, once created, will always produce a
valid status property is incorrect. Instead, expired orders may return
404s.

Handle this by including status code information in errors returned from
the ACME client. Handle this in 'shaman' by modifying the 'domain' block
in such case.

Additionally, erase the 'order_id' property and replace it with the
'certificate_id' property in a 'domain' block once the order has been
finalized.

This is a backwards incompatible change due to the reinterpretation of
fields in 'shaman.json'.