concorde.client: encode uints with fixed sizes
Fix the encoding of JWA signatures to include leading zero octets, as
per the ACME spec.

An ECDSA signature consists of a pair of integers ranging from one to
one less than the order of the base point of the elliptic curve (this is
also the range of the private key).

As a result, the encoding of one of these integers as big-endian octets
can have varying lengths due to leading zero octets. They will always
have a maximum length that will also correspond to the size of the key.

The ACME (and underlying JWA) spec mandates that all leading zero
octets are included (rounding up to a whole number of octets) for ECDSA
signatures. See https://tools.ietf.org/html/rfc7518#section-3.4 for more
information.

Take this opportunity to rename some of the methods in `concorde.acme`.